+++ fabric/src/main/java/cc/cassian/item_descriptions/client/limelight/LimelightIntegration.java
@@ -1,7 +1,7 @@
 package cc.cassian.item_descriptions.client.limelight;
 
-//? if 1.21.1 {
-/*import io.wispforest.limelight.api.LimelightEntrypoint;
+//? if 1.21.1 || 1.21.5 {
+import io.wispforest.limelight.api.LimelightEntrypoint;
 import io.wispforest.limelight.api.extension.LimelightExtension;
 
 import java.util.function.Consumer;
@@ -13,4 +13,4 @@ public class LimelightIntegration implements LimelightEntrypoint {
         extensionRegistry.accept(DescriptionsExtension.INSTANCE);
     }
 }
-*///?}
+//?}