+++ src/main/java/cc/cassian/item_descriptions/client/mixin/StatusEffectsDisplayMixin.java
@@ -18,6 +18,10 @@ import net.minecraft.item.tooltip.TooltipData;
 //? if >1.20 {
 import net.minecraft.client.gui.DrawContext;
  //?}
+//? if =1.19.2 {
+/*import net.minecraft.client.util.math.MatrixStack;
+*///?}
+import net.minecraft.entity.effect.StatusEffectInstance;
 import net.minecraft.text.Text;
 import org.spongepowered.asm.mixin.Mixin;
 import org.spongepowered.asm.mixin.Shadow;
@@ -34,10 +38,21 @@ import java.util.Optional;
 //?} else
 /*@Mixin(AbstractInventoryScreen.class)*/
 public abstract class StatusEffectsDisplayMixin {
+
+    //? if =1.21.5 || =1.21.4 {
     @Shadow protected abstract void drawStatusEffectDescriptions(DrawContext context, int x, int height, Iterable<StatusEffectInstance> statusEffects);
 
+    @Inject(method = "drawStatusEffects(Lnet/minecraft/client/gui/DrawContext;II)V", at = @At(value = "INVOKE", shift = At.Shift.AFTER, target = "Lnet/minecraft/client/gui/screen/ingame/StatusEffectsDisplay;drawStatusEffectSprites(Lnet/minecraft/client/gui/DrawContext;IILjava/lang/Iterable;Z)V"))
+    private void forceShowDescriptions(DrawContext context, int mouseX, int mouseY, CallbackInfo ci, @Local LocalBooleanRef booleanRef, @Local(name = "i") int i, @Local(name = "k") int k, @Local Iterable<StatusEffectInstance> iterable) {
+        this.drawStatusEffectDescriptions(context, i, k, iterable);
+        booleanRef.set(false);
+    }
+
+    //?}
     //? if >1.21.5 {
-    /*@Redirect(method = "drawStatusEffectTooltip", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/DrawContext;drawTooltip(Lnet/minecraft/client/font/TextRenderer;Ljava/util/List;Ljava/util/Optional;II)V"))
+    /*@Shadow protected abstract void drawStatusEffectDescriptions(DrawContext context, int x, int height, Iterable<StatusEffectInstance> statusEffects);
+
+    @Redirect(method = "drawStatusEffectTooltip", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/DrawContext;drawTooltip(Lnet/minecraft/client/font/TextRenderer;Ljava/util/List;Ljava/util/Optional;II)V"))
     private void renderEffects(DrawContext context, TextRenderer textRenderer, List<Text> text, Optional<TooltipData> data, int mouseX, int mouseY) {
         List<Text> tooltipText = ModHelpers.createEffectDescription(text);
         context.drawTooltip(textRenderer, tooltipText, Optional.empty(), mouseX, mouseY);
@@ -60,11 +75,26 @@ public abstract class StatusEffectsDisplayMixin {
         List<Text> tooltipText = ModHelpers.createEffectDescription(text);
         context.drawTooltip(textRenderer, tooltipText, Optional.empty(), mouseX, mouseY);
     }
+    @Shadow protected abstract void drawStatusEffectDescriptions(DrawContext context, int x, int height, Iterable<StatusEffectInstance> statusEffects);
+
+    @Inject(method = "drawStatusEffects(Lnet/minecraft/client/gui/DrawContext;II)V", at = @At(value = "INVOKE", shift = At.Shift.AFTER, target = "Lnet/minecraft/client/gui/screen/ingame/AbstractInventoryScreen;drawStatusEffectSprites(Lnet/minecraft/client/gui/DrawContext;IILjava/lang/Iterable;Z)V"))
+    private void forceShowDescriptions(DrawContext context, int mouseX, int mouseY, CallbackInfo ci, @Local LocalBooleanRef booleanRef, @Local(name = "i") int i, @Local(name = "k") int k, @Local Iterable<StatusEffectInstance> iterable) {
+        this.drawStatusEffectDescriptions(context, i, k, iterable);
+        booleanRef.set(false);
+    }
     *///?} else {
-    /*@Redirect(method = "drawStatusEffects", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screen/ingame/AbstractInventoryScreen;renderTooltip(Lnet/minecraft/client/util/math/MatrixStack;Ljava/util/List;Ljava/util/Optional;II)V"))
+    /*@Shadow protected abstract void drawStatusEffectDescriptions(MatrixStack matrices, int x, int height, Iterable<StatusEffectInstance> statusEffects);
+
+    @Redirect(method = "drawStatusEffects", at = @At(value = "INVOKE", target = "Lnet/minecraft/client/gui/screen/ingame/AbstractInventoryScreen;renderTooltip(Lnet/minecraft/client/util/math/MatrixStack;Ljava/util/List;Ljava/util/Optional;II)V"))
     private void renderEffects(AbstractInventoryScreen instance, MatrixStack stack, List<Text> text, Optional optional, int mouseX, int mouseY) {
         List<Text> tooltipText = ModHelpers.createEffectDescription(text);
         instance.renderTooltip(stack, tooltipText, Optional.empty(), mouseX, mouseY);
+    }
+
+    @Inject(method = "drawStatusEffects", at = @At(value = "INVOKE", shift = At.Shift.AFTER, target = "Lnet/minecraft/client/gui/screen/ingame/AbstractInventoryScreen;drawStatusEffectSprites(Lnet/minecraft/client/util/math/MatrixStack;IILjava/lang/Iterable;Z)V"))
+    private void forceShowDescriptions(MatrixStack matrices, int mouseX, int mouseY, CallbackInfo ci, @Local LocalBooleanRef booleanRef, @Local(name = "i") int i, @Local(name = "k") int k, @Local Iterable<StatusEffectInstance> iterable) {
+        this.drawStatusEffectDescriptions(matrices, i, k, iterable);
+        booleanRef.set(false);
     }
     *///?}